+++ src/_lib/startOfUTCWeek/test.js
@@ -115,4 +115,8 @@ describe('startOfUTCWeek', function () {
     var block = startOfUTCWeek.bind(null, date, {additionalDigits: NaN})
     assert.throws(block, RangeError)
   })
+
+  it('throws TypeError exception if passed less than 1 argument', function () {
+    assert.throws(startOfUTCWeek.bind(null), TypeError)
+  })
 })