+++ src/fp/getISOYearWithOptions/index.js
@@ -1,8 +1,0 @@
-// This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
-
-import fn from '../../getISOYear/index.js'
-import convertToFP from '../_lib/convertToFP/index.js'
-
-var getISOYearWithOptions = convertToFP(fn, 2)
-
-export default getISOYearWithOptions