+++ src/locale/it/_lib/formatDistance/index.js
@@ -57,8 +57,8 @@ var formatDistanceLocale = {
   },
 
   overXYears: {
-    one: 'piú di un anno',
-    other: 'piú di {{count}} anni'
+    one: 'più di un anno',
+    other: 'più di {{count}} anni'
   },
 
   almostXYears: {
@@ -81,7 +81,7 @@ export default function formatDistance (token, count, options) {
 
   if (options.addSuffix) {
     if (options.comparison > 0) {
-      return 'in ' + result
+      return 'tra ' + result
     } else {
       return result + ' fa'
     }